requirements.txt
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
fonttools==4.54.1
h5py==3.12.1
huggingface-hub==0.16.4
idna==3.10
ipython==8.12.3
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kiwisolver==1.4.7
markupsafe==3.0.1
matplotlib==3.5.3
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.23.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==9.5.0
pipreqs==0.5.0
prompt-toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sacremoses==0.1.1
scikit-learn==1.0.2
scipy==1.13.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
staintools==2.1.2
threadpoolctl==3.5.0
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.12.1
torchstain==1.3.0
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.30.2
urllib3==2.2.3
wandb==0.18.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
einops==0.8.0
huggingface_hub==0.16.4
kornia==0.7.3
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
grad-cam==1.5.4
scikit-image==0.24.0